Move sudokuGame logic back into custom hook
